Change CLI Flags to all use the same format
The flags provided use different syntaxes, most are `--flag=value` while some are `--flag value`. 
It shouldn't really make a difference when run, but im trying to debug an issue with making calls to the beacon API this was one thing i noticed, and am changing to all be `--flag=value` for stylistic purposes if nothing else.